Moved the skipped prompt logic to the source.
